Introduce RemoteOp API
Change Homa interface to be invoking and receiving remote operations
(RemoteOp).  This matches a request with a response like in RPC but
allows the request to be delegated or forwarded and for the response
to come from the delegated server.  Introduces/modifies 3 concepts:

    1)  A Message is an array of bytes that form either a request or
         response.
    2)  A RemoteOp is a request/operation to be processed on a
        remote server.
    3)  A ServerOp is a server-side handle of an incoming RemoteOp.

A refactoring of the internal implementation will follow.