Cleanup Message Interface and Implementation
Add Message header awareness to the Message class so that the
Transport can easily store and retrieve the message header.
Remove the set() method which forces the class to support cases
where there are "holes" in the Message. Removing set() allows
the class to make the simplifying assumption that all data
between the beginning and end of the Message is valid.