Move OpContext handling into Receiver
Move some of the logic of handing inbound messages back into the
Receiver.  Previously, some of this was pushed up into the Transport
but it didn't result in a good separation of responsibilities.  Now
the Transport should be responsible for managing and understanding
RemoteOp/ServerOp concepts (which ties messages together).  The
Receiver on the other hand will manage Message level concerns.