Move OpContext handling into Sender
Move some of the logic of handing outbound messages back into the
Sender.  Previously, some of this was pushed up into the Transport
but it didn't result in a good separation of responsibilities.  Now
the Transport should be responsible for managing and understanding
RemoteOp/ServerOp concepts (which ties messages together).  The
Sender on the other hand will manage Message level concerns.