Optimize poll loop
The Transport::poll method must contain some amount of code that
executes even when the poll loop performs no useful work (e.g. to
check if any work needs to be done).  This commit improves the
performance of various components of the poll loop when there is
no useful work to do.