Change default behavior to ensure message is sent
Homa will now ensure that an OutMessage is fully sent before it is
cleaned up if an application releases the message after calling
send(). Previously, released messages that were in progress of
being sent were cancelled and immediately cleaned up. Canceling a
message now requires an explicit call to cancel().