Remove Sender::trySend() restriction
Previously, the number of messages completed per call to trySend()
was limited by an std::array allocated on the stack. This was done
to reduce the allocation overhead. New Perf microbenchmarks show
that reserving capacity in an std::vector is likely efficient
enough and would allow trySend() to seamlessly handle an unlimited
number of message completions per call to trySend().